feat: introduce logion-specific config.
logion-network/logion-internal#1283
